Encode main without return rule.
If main does not have return, then return 0 must be generated.
Put this in a separate file, calling it a codegen rule.
The existing check is not sufficient - if main ends in a conditional,
it will still append a return 0 (as it explicitly looks for Return at
the top level of the function's AST).